Allow sliding the mouse over the piano keys
Since Qt happens to grab the mouse when you click a widget, all mouse events will only go to that widget until you release the mouse. So in order to allow sliding the mouse over the piano keys, we needed a "hacky" solution that involves detecting mouse move + left button pressed events within every key, figuring out which key the mouse is hovering over, then sending a signal for that key to be pressed. This solution was adapted from C++ code shown here: https://forum.qt.io/topic/107430/detect-if-mouse-buttons-are-pressed-when-sliding-into-widget/16