Implement dragging to make a selection box
The QRubberBand() was a temporary solution before I found out QGraphicsView has a built-in rubberband selection feature, accessible by using setDragMode(). As of this commit, *two* rectangles are drawn when you drag the mouse across the note block area (one is the built-in feature, and the other was implemented manually according to https://doc.qt.io/qt-5/qrubberband.html, which is meant for normal widgets instead of QGraphicsItem's). I'm only committing this for future reference, since, due to some particularities of the dragging modes, I *might* have to implement the dragging mechanic myself. We'll see.

Also, pointy hands when you hover the piano and the note block area, as in the original version.