Implement custom selection dragging for note block area
So, it turned out exactly as I was expecting in 43e2524 - the built-in rubber band in QGraphicsScene is useful but very limited, neither allowing automatic scroll when you reach the edges of the view, nor to change its color/style - let alone filter out when you want it to appear vs. when you don't (i.e. if you want to require pressing a key on the keyboard to make a selection or if you only want it to happen if you hold the left mouse button).

Trying to implement a separate QRubberBand yourself will result in the weirdest behaviors, ranging from it drawing outside the main window (even across different monitors), not updating its position when you scroll the view (i.e. the top left corner will not stay attached to where you first clicked but rather scroll together with the view), OR updating it incorrectly (i.e. instead of the top position scrolling up when you scroll the view down, it remains at the same position on the screen and the rectangle descends *below* the mouse position while you're dragging it). See: http://web.archive.org/web/20180222121837/http://www.qtcentre.org/threads/10741-Rubberband-relative-to-QGraphicsScene

I tried to set the view as the rubber band's parent but that turned out not to work, so I simply added it to the scene via addWidget(). This method returns a QGraphicsProxyItem (a graphic item that embeds a widget into a scene) that then has its Z value set to 1, in order to draw itself on top of the other items (otherwise the selection box crawls under the note blocks and gridlines as you drag it). See: https://stackoverflow.com/questions/24282371/adding-a-qwidget-to-a-qgraphicsscene

Since we reimplemented the mouseMoveEvent for the scene, note blocks were no longer able to detect when you hovered over them. So when no button is pressed as you move the mouse, the event handler will call the parent's (QGraphicsScene) method, which in turn will forward the event to the note blocks as it would have done normally. See: https://stackoverflow.com/questions/2940392/qgraphicsitem-doesnt-receive-mouse-hover-events

Additional useful links:
https://stackoverflow.com/questions/46113608/bad-rendering-of-qrubberband-in-qgraphicsscene
https://stackoverflow.com/questions/42692885/qgraphicsview-how-to-make-rubber-band-selection-appear-only-on-left-mouse-butto