Prevent moving note blocks outside of the view
This is sort of inefficient because it has to iterate the whole selection three times. Preferably I'd use some sort of bounding box function to figure out if any blocks are out of bounds, but Qt provides no such thing for selected items, only for the whole view. So that's the best we have at the moment...