Change addBlock and removeBlock methods to use grid coordinates
There are more situations where you would use the "tick" and "layer" properties when creating a note block than its position inside the scene. It's the method's job to figure out where that tick and layer will lie inside the scene. With that, right now we have the following scheme:

1. Get the location of the click with event.scenePos()
2. Find the grid coordinates that this point is contained within through getGridPos()
3. (inside each method) Find the top-left corner of that grid position through getScenePos() and place the note block there.