Enable cache on note blocks to cut down drawing time
Enabling item cache provides an immense performance gain, especially when zoomed out and rendering many note blocks at once. While the performance with thousands of items is still not optimal, there's still a lot we can do to optimize it - for instance, reducing the level of detail when zoomed out very far.