Move ruler widget to NoteBlockView
Syncing the marker head to the body is a *terrible* idea. This change is so the ruler, rather than being a standalone widget that's independent of the NoteBlockView, is added as a child widget of it, and positioned in its top margin (it's now embedded inside the NoteBlockView). This allows us to draw the marker as a single QWidget spanning both the ruler and the note block area itself.