Change LayerBar to be a QFrame instead of a QToolBar
This caused the wrong widget hierarchy -- since the QToolBar is contained inside a QFrame, having the latter as an attribute of the former required manipulating `layer.frame` instead `layer` directly, which is a really bad idea. The QToolBar is now an attribute of the QFrame, which makes much more sense.