Disable actions according to the scene being empty
The empty status being tied to `updateSelectionStatus` creates a weird requirement, since placing or removing a note block doesn't emit the `updateSelectionStatus` signal, but can make the scene become empty or non-empty. As such, this state is now handled in a separate signal chain that's updated according to the `blockCountChanged` signal (which can later be reused for e.g. displaying the total note count in the status bar).