Add action managers for instrument-related actions
Some actions, like changing the current instrument, have to be accessed from multiple places (e.g. Settings menu and toolbar). So it doesn't make sense for them to be stored on the menu they belong to, but rather, in a specialized class that handles the creation and updating of actions related to instruments.

Clients can then subscribe to signals sent from the manager, as well as emit signals for its state to be altered via slots.

We have chosen to use a module-level variable to store

, since it can be easily imported by other components instead of passed around everywhere. This is not a very elegant soution as it requires the `global` keyword so that the manager classes can update the module-level action lists, but it lays the foundation for a pattern I'd like to use more in the application, since the concept of a manager seems appropriate for use in a context like this - it could just use some refinement in the future.