Use dependency injection for central area widgets
It's better for the main window to instantiate the workspace objects, because this centralizes the responsibility to handle those widgets, removing the need to peek inside three levels of inner variables to access the note block area and getting direct access to the signals contained in each one, besides making testing easier.

The same can be done further with the inner note block area components (ruler, marker, view etc.), if it's shown to have a clear advantage.