Fix clicked instrument buttons getting the wrong ID
Previously, all buttons where sending that id 15 (last one) was pressed (for some reason the lambda with the ID was not working, despite being the same thing we did with the `QActions`. A `QButtonGroup` solved this as it has a single signal `idClicked` with the ID of the button that was clicked (and also fixes the problems of multiple orphan connections when we re-create the buttons).