Play current instrument when pressing piano key
We've also moved initialization of the instrument managers to the `__init__` method (which ideally should be the only one to declare new parameters on the `MainWindow`), since the piano relies on it to know what instrument is currently selected.

By the way, I'm trying a new state lookup approach here - I was about to add a 'currentInstrument' property to the piano itself, so it can keep track of it when you change an instrument to play the correct one. But instead, I've figured I could just add a `currentInstrument` property to the `InstrumentManager` (since it belongs more there), because we only need to know what instrument is currently selected by the time we play a note. So it makes much more sense for the `MainWindow` to look that up at the signal connection than it does to update the piano and have *it* send the current instrument ("Why should *I* keep care of *your* instruments?? I don't care about them!!)

This will probably help isolate responsibilities a little bit. I already feel like I'm passing too many signals around...