Move `PlaybackManager` to its own module and make it top-level
By lifting up the `PlaybackManager` to be managed by the `MainWindow`, it's now completely transparent _where_ playback is handled inside the application, and who's able to control parameters regarding playback -- whereas, before, connections to the `PlaybackManager` where scattered across three different hierarchical levels (`Workspace`, `NoteBlockView` and each of its subwidgets).

This is exactly the kind of thing I'd like to do a lot in this codebase -- less shared responsibilities, more 'central' locations for doing one thing and doing it well. As such, we can think of the `ui` folder as the view, `controller` as controller, and `core` as the model -- a bit loosely due to the Qt's Model-View paradigm, which is a slightly modified version of MVC, and due to mixing native Python features with Qt's offerings.

This is one of the best refactors I've done in my entire life. =)