Remove rubber band selection from being a scene item
When I had originally implemented the rubber band selection, simply making it a child of the `QGraphicsView` caused its initial position not to be anchored to the scene, meaning it would stay in the same place as the view was scrolled. As a result, I've made the rubber band a scene item, sharing its space with note blocks themselves. This, however, requires a lot of shenanigans when dealing with the scene items, since there's always ONE exception -- one item that is not a note block, and, as such, can't be returned when you fetch the note blocks, can't be counted when you want the note count etc.

That subsequent problem was the motivation behind implementing `noteBlocks()` (which removes all the non-note-block items from the result), which not only creates an unnecessary complexity layer everytime we need to manipulate all note blocks in the scene, but was also implemented extremely unneficiently, to make matters worse (it iterates all scene items and yields it only if it's an instance of a note block, which is true for `n-1` scenarios where `n` is the amount of items on the scene). Another approach could be to get the entire list of items and `pop()` the selection item before returning it, but that would still incur on the performance penalty of handling the list in Python, which is really not ideal - instead, it's a good idea keep things to be handled by the Qt C++ code, whenever possible.

As it turns out, that was just a solution to the wrong problem. The proper solution, implemented in this commit, was to make the rubber band selection a child of `QGraphicsView` (as I had originally done two years ago, but gave up too early on trying to make it work), and map the view coordinates to scene coordinates in `mouseMoveEvent`, so it follows the original click position as the view is scrolled.