Use native Qt's implementation for `setSelectionArea`
More efficient, as it avoids iterating every note block on the Python end in favor of using Qt's optimized operations.

As there's no way to subtract a selection from the selection area (the only two possible item selection operations are `ReplaceSelection` (default) and `AddToSelection`), we use Qt's native implementation for adding a region to the selection, which is done much more often, and fall back to removing blocks on an individual basis when that is the case.