Use a single timer to control glowing effect of all notes
Running a `QTimer()` on every playing note block is a shot in the foot. Keeping track of the glowing notes, then updating them from outside with a global timer puts much less stress on Qt's event system (`NoteBlock` has to be as "dumb" as possible since we can have thousands of them shown on screen at the same time).