Replace `@property` with plain `update()` calls for opacity
Properties add quite a bit of overhead (and code). As per the link below, it's not advisable to use a `property` when you're only interested in rewriting the setter procedure.
https://stackoverflow.com/questions/17576009/python-class-property-use-setter-but-evade-getter

Instead, one would be better off using `__setattr__` (which adds much more overhead) or... simply going the non-pythonic route and writing a setter method. *facepalm*

I believe it can be forgiven here since we need to squeeze every last bit of performance from the note block update logic. Also, this is just what you would do by calling `update()` or `repaint()` as one does when using PyQt (it's not a very pythonic binding anyway).

Another reason is that we don't have a client setting opacity externally to need such encapsulation - the entire control over each item's opacity is on the item itself. There's no need to hide implementation details from the class itself(!)