Move selected overlay drawing outside of cache
The point of the cache is reusing as much painting as possible, so it doesn't make sense to redraw the entire note block when only the selection overlay is supposed to change. As such, we can draw the note block from the cache, and then simply draw the overlay on top of it if need be.