Rewrite audio engine to use real-time pitch shifting
Using `pydub` to generate the entire resampled sound into memory at playback time is really, REALLY not a good idea (internally, it uses the built-in python `array` for performing audio manipulation, which is really slow and in no way intended for real-time audio manipulation).

The new algorithm uses `python-samplerate`, which implements a binding to `libsamplerate` (a.k.a. Secret Rabbit Code) to resample the audio in real time, fetching samples as requested by the audio output stream via a callback function. This solution is much more appropriate, as it resamples the audio lazily instead of immediately (i.e. it only generates the samples that are needed at a given time from each currently played sound, putting much less CPU strain).

`pydub` is still used, but only to decode the sound as it's loaded for the first time. As a result, this means that audio samples are internally only handled using `numpy` arrays. This is also a huge speed gain, as it avoids the conversion step from `array` to `ndarray` so the samples can be sent to the output stream.

And the best part: as we're bypassing pydub completely, we must also implement panning and volume changes, which was really easy - changing the volume is as simple as multiplying the numpy array by the volume factor, and panning is basically the same thing, only with a bit more calculation involved to figure out the L and R gain factors. The bonus of handling everything ourselves is that, in case we REALLY care about micro-optimizations in the future, the volume and pitch of the sound can be set in one go, which avoids one array multiplication operation entirely.