Implement audio resampler object pool
Creating a new instance of `CallbackResampler` for every played sound was taking most of the CPU time for playing a sound, since `sounddevice` has to call ffi functions to register the callback function.

With the object pool pattern, it's possible to generate only the amount of resamplers needed for concurrent playback, and allocate them to different sounds as needed, avoiding the initialization time entirely and only replacing it with `reset()` (which is much faster).