Remove resampling via callbacks and do it all it once
You know the thing that goes "Do not become attached to your code"? Yeah... the thing I just spent three days on absolutely doesn't work! It's way too costly to iterate on 256 different resamplers, all trying to do calculations in the short timeframe the output callback has to be returned from. With only a few sounds at once, it was already impossible to return from the function on time, so output underflows were almost guaranteed under almost no load on the audio engine.

At first, I was wondering if the problem was with numpy not being able to crunch the samples fast enough. Where numpy really shines is on bypassing pure python operations, and we weren't doing that - the `for` loop in the playback callback iterating every sound instance was "chopping" all the really fast numpy operations into multiple slices, still making things slow. This turned out not to be the problem.

Turns out there's a bit of overhead when calling the resampler for real-time processing, since it has to reach the low-level stuff everytime you request new samples. When you're trying to do that 256 times in under 15 milliseconds (approximate time between each `input_callback` call), it becomes almost impossible to process samples in enough time.

Sometimes, the simplest (and obvious) approach can fix everything. Before turning to `libsamplerate`, sample rate conversion was being done by `pydub`, which operates in python arrays and is really slow. In e14e94595de31f8261e1e95b31d0e857e15f7320, real-time sampling was implemented *because* `libsamplerate` had been introduced, and it offered such a capability (which seemed really appropriate at first, so... why not go fancy?). The mediocre performance was followed by fcdce196189614db77e4b4d4fdff8836d4e4cdde, which improved performance but not by much. It turns out I hadn't tried the obvious solution -- replacing `pydub` with `libsamplerate`, but still doing all processing immediately. As `libsamplerate` is much, MUCH faster, it's able to keep up with resampling hundreds of sounds while barely causing any slowdown to the program. We basically removed the "box" around the processing time caused by resampling inside the `callback` function itself, and allowed it to take as much time as it needs.

As the processing time is now split with UI rendering time, it does slow things down a little bit when *too many* sounds are playing. The audio itself, on the other hand, works fine -- which is arguably the most important aspect in an audio application, so a bit of slowdown in the UI in favor of precise playback may be an acceptable trade-off. Nevertheless, there are still optimizations we can make -- for instance, adding a short-term cache to played sounds to avoid resampling previously resampled sounds, and save a bit on the CPU load. The ideal approach might also be a middle ground between both solutions: call the resampler every n callback calls so it dumps more samples into a buffer, then have a few calls without having to call it.