Implement resampling cache
In any ordinary song, there's usually a small pool of notes that are repeated more commonly, due to songs being written in a certain key or using only a small set of instruments. By implementing a very simple cache for resampled sounds, we can avoid re-processing notes that have previously been played, for a very neat performance boost (dict lookup is O(1), so it's very fast to retrieve previously generated samples from the cache).

This, of course, bumps up the memory requirements. Currently, the cache is unlimited, which may lead to very high memory usage given enough use of the program. But if a tighter memory constraint is required, this simple idea can be further improved upon, by e.g.

- resetting the cache when a new song is loaded;
- pre-loading the most used sounds in the song, or doing it asynchronously in idle CPU time;
- setting a limit to the number of cache entries, and clearing the n oldest entries;
- clearing the least used entries, as they're less likely to be used again.