Fix changing volume of played samples affecting original sound
As slicing a numpy array returns a view of the array, any modifications done to it (such as applying volume and panning) affects the original sample data. This was not a problem before due to resampling all sounds, which returned an entirely new array. With the implementation of the cache, though, the original sample array is now being passed to `SoundInstance`, which causes the original data to be affected when volume and panning are applied. To prevent this, we need to take a copy of the array when retrieving it from the cache (which is a small cost compared to resampling it entirely).