Move `AudioEngine` to separate thread
Audio playback is still done in a blocking manner, since it's CPU-bound. However, loading sounds is now non-blocking, so the application loads immediately while the default instrument sounds are loaded in the background. The audio engine is now also shut down properly when the application ends.

There might be a way to also resample/play sounds in a multi-threaded way, but this could introduce issues since the application has to *constantly* be able to push new sounds to the engine.