refactor(layer): defer handling of layers to `LayerController`
tl;dr: This is a major refactoring to reduce coupling, improve the single responsibility principle and make it easier to modify layer data from external places.

---

Prior to this commit, handling layers was the shared responsibility of `LayerArea` and `NoteBlockArea`: the former created and emitted new layers, while the latter kept hold of the layer data (it only made sense, since the `NoteBlockArea` uses layer data extensively to play notes, show the locked overlay etc.)

However, this created a weird situation where both parties had the responsibility to keep the data in sync, and the outside world had no "easy" access to that data.

This could be solved by syncing their state in the `Workspace` and only communicating with it from the "outer world" (it would make sure the shared state was always in sync). While this could solve the problem, it still doesn't make much sense to store the layer data in the `NoteBlockWidget` - it only existed there as a matter of convenience, but layer data doesn't pertain to note blocks in any way.

Since there was not a central location to serve as an "emitter" of layers, `LayerController` was introduced, which holds the business logic behind creating, editing and removing layers. This decreases coupling between different components, allowing every client who's interested in reading or managing the layer data to subscribe to a single central location (single source of truth). As a result, song loading can now interact with `LayerController` to notify all client classes about the layers that should be created. More importantly, it will allow the plugin API to deal with layers without needing to introspect data that's held three layers deep inside a deeply nested widget (which it should not be aware exists).

As a consequence, a few bits in some components' external interfaces were changed:

- `NoteBlockArea.updateSceneSize()` was changed to return grid coordinates (1, 2) instead of pixel coordinates (32, 64), which is a better choice anyway (the entire outside world only cares about layer/tick positions, not pixels -- they're a metric internal to `NoteBlockArea` and the workspace!) This allows the synchronization to happen in `MainWindow` without it being aware of the `BLOCK_SIZE` constant.

- `NoteBlockArea` no longer keeps an internal list of layers, but receives one as an argument. This is crucial to the single responsibility principle -- it's layer handling methods now only update the scene data, not the global list of layers, yet it still has access to the layer data.

- A `Layer` dataclass was added to to the `nbs.core.data` module, and used in the layer UI module instead of the data in `nbs.core.song`.

- Signals in the `LayerArea` class were changed from `___Changed` to `___ChangeRequested`, since they doesn't effectively change the *actual* layer data anymore.

- The layer count is no longer managed by the `LayerArea`/`NoteBlockArea`, but by `LayerManager`. As such, the `updateLayerCount` methods were removed from both.

- The index and layer data are now required arguments to the `LayerArea`'s `addLayer()` method, because it no longer handles the creation of layers in case you don't pass one. Similarly, it also no longer emits signals on changes to layers, but only on changes to the control widgets -- it's now "passive" and acts more as a listener.

- Individual layers now have proper methods to update their data externally. This is to allow loading layer data from elsewhere.