fix(layer): shrink workspace immediately when min layer count decreases
When there's a populated layer holding the workspace from shrinking, as soon as it becomes an empty layer (i.e. you unlock it, or delete the last character of its name), the workspace will collapse to the next minimum (i.e. either the last populated layer or the last "fake" layer created for margin).