refactor(playback): make `PlaybackController` single source of truth
Part of a larger refactor for responsibility inversion: the song tempo is no longer managed by the time bar, nor the tempo box; it now merely "requests" the `PlaybackController` to change that information, which sends back the acknowledgement so that it can update itself. Also removes direct connection between the `NoteBlockView` and the `TimeBar` (they have no business talking to each other!).

This allows the tempo to be changed by external sources (such as when loading a song, using the future plugin API, making the song compatible etc.) without directly interacting with the tempo widget. Additionally, the tempo may now be displayed in other places while only needing to "subscribe" to a single place.