perf(workspace): implement ruler paint cache
Avoids redrawing the entire time ruler when any part of it gets revealed. Drawing lots of elements, especially text (whose bounding box is expensive to calculate) takes time, and even if we reduced the drawing done to only comprise the updated rect received by the `paintEvent`, any piece of text covered would still have to be redrawn completely. The implemented solution still requires a full redraw every now and then to update the cache, but once that's done, hundreds of subsequent redraws can be avoided completely.