perf(note): replace `QGraphicsOpacityEffect` with `QPropertyAnimation`
`QGraphicsOpacityEffect` has no built-in interpolation/animation, requiring us to manually iterate through every playing note block in a timer, decreasing the opacity. This, naturally, anniquilates any sign of good performance during playback.

Qt, on the other hand, offers `QGraphicsItemAnimation` for applying animations to `QGraphicsItem`s. Unfortunately, it can only animate item transforms (rotation, translation, scale) - no such luck if we want to animate the opacity.

The last resort is to go with `QPropertyAnimation`, which can animate a variety of widget properties (including opacity), but unfortunately, only operates in `QObject`s. The Qt docs recommend using multi-inheritance from `QObject` and `QGraphicsItem` (see: https://doc.qt.io/qt-6/graphicsview.html#animation ), which has less of an overhead than making it a `QGraphicsObject` or just straight using `QObject`s (since many unnecessary signals will be emitted all the time, which becomes a problem when you have hundreds of thousands of items). This approach can reduce the overhead of emitting unnecessary signals such as `itemChanged` (which is the point of using `QGraphicsItem`s in the first place).

Sadly (what a rollercoaster!), PyQt has no proper support for multiple inheritance from multiple Qt classes. The final solution is to make the `NoteBlock` class a `QGraphicsObject` instead of a `QGraphicsItem`. This has the immediate consequence of increasing song load times as the time per `__init__` call on each note added to the song increases dramatically. It's really a waste to go with a full-fledged `QObject`, signals- and slots-enabled, when all we need is to animate some opacities. :/

I guess it isn't too bad after all - the performance hit of letting Qt manage the animation updates itself is much lighter!